ShoppingList: passing intents to MainActivity done

+++ ShoppingList/app/src/main/java/io/github/soojison/shoppinglist/data/Category.java
@@ -1,10 +1,10 @@
 package io.github.soojison.shoppinglist.data;
 
 public class Category {
-    public static final short FOOD = 1;
-    public static final short TOILETRIES = 2;
-    public static final short EDUCATION = 3;
-    public static final short APPAREL = 4;
-    public static final short ENTERTAINMENT = 5;
-    public static final short MISC = 6;
+    public static final int FOOD = 1;
+    public static final int TOILETRIES = 2;
+    public static final int EDUCATION = 3;
+    public static final int APPAREL = 4;
+    public static final int ENTERTAINMENT = 5;
+    public static final int MISC = 6;
 }
